Move compression related steps from createrepo_c to common
