+++ dnf-behave-tests/createrepo_c/steps/lib/file.py
@@ -26,14 +26,6 @@ def get_checksum_regex(type_str):
     raise ValueError("Unknown checksum type: " + type_str)
 
 
-def get_compression_suffix(type_str):
-    if type_str in ("gz", "zck", "xz", "bz2"):
-        return "." + type_str
-    if type_str == "-":
-        return ""
-    raise ValueError("Unknown compression type: " + type_str)
-
-
 def decompression_iter(filepath, compression_type, blocksize=65536):
     if compression_type == "gz":
         return file_as_blockiter(gzip.open(filepath, 'rb'), blocksize)
@@ -84,21 +76,3 @@ def decompress_file_by_extension_to_dir(compressed_filepath, dest_dir):
 
     open(dst, "wb").write(content)
     return dst
-
-
-def create_compressed_file_with_contents(filename, compression, contents, encoding="utf-8"):
-    fullname = filename + get_compression_suffix(compression)
-    if os.path.exists(fullname):
-        raise ValueError("File: " + fullname + " already exists")
-
-    if compression == "gz":
-        with gzip.open(fullname, 'wt') as f:
-            f.write(contents)
-    elif compression == "xz":
-        with lzma.open(fullname, 'wt') as f:
-            f.write(contents)
-    elif compression == "bz2":
-        with bz2.open(fullname, 'wt') as f:
-            f.write(contents)
-    else:
-        raise ValueError("Unknown compression type: " + compression)
